+++ packages/fluent-formio/CHANGELOG.md
@@ -1,5 +1,11 @@
 # 0.5.20
 
+## 0.8.7
+
+### Patch Changes
+
+- @goatlab/fluent@0.9.7
+
 ## 0.8.6
 
 ### Patch Changes